ScenegraphLayer: Final round of updates
