+++ examples/layer-browser/package.json
@@ -9,7 +9,8 @@
     "start-local-production": "webpack-dev-server --env.local --env.production --progress --hot --open"
   },
   "dependencies": {
-    "@loaders.gl/gltf": "^0.6.1",
+    "@loaders.gl/core": ">=1.0.0-alpha.2",
+    "@luma.gl/addons": "^7.0.0-beta",
     "@luma.gl/debug": "^7.0.0-alpha",
     "colorbrewer": "^1.0.0",
     "d3-request": "^1.0.6",
